requirements.txt
flask
flask-cors
gTTS
google-generativeai
python-dotenv
